notifications/requirements.txt
mysql-connector-python==8.0.22
twisted==18.7.0